feat: Implemented method to update the values of each trainable with SGD
Stochastic Gradient Descent (SGD) is a version of Gradient Descent where on each forward pass a batch of data is randomly sampled from total dataset. SGD is an approximation of Gradient Descent, the more batches processed by the neural network, the better the approximation. Implemented method to update the values of each trainable with SGD.