feat: Added sentiment classification Neural Network
Created a basic neural network much like the networks you've seen in earlier lessons and in Project 1, with an input layer, a hidden layer, and an output layer.
Not addded a non-linearity in the hidden layer. That is, not used an activation function when calculating the hidden layer outputs.
Re-used the code from earlier in this notebook to create the training data
Implemented the pre_process_data function to create the vocabulary for our training data generating functions