feat: Reducing Noise by Strategically Reducing the Vocabulary
Introduced two more parameters min_count and polarity_cutoff for reducing noise. Changed so that words are only added to the vocabulary if they occur in the vocabulary more than min_count times. Changed so that words are only added to the vocabulary if the absolute value of their postive-to-negative ratio is at least polarity_cutoff.