feat: Handwritten Number Recognition with TFLearn and MNIST
In this notebook, we have be built a neural network that recognizes handwritten numbers 0-9.

This kind of neural network is used in a variety of real-world applications including: recognizing phone numbers and sorting postal mail by address. For building this network, we have used the MNIST data set, which consists of images of handwritten numbers and their correct labels 0-9.